거래내역 지출,수입 저장 로직 수정
