+++ db.json
@@ -67,7 +67,8 @@
       "incomeCategory": "급여",
       "tradeDate": "2025-03-01",
       "tradeDescription": "3월 급여 입금",
-      "userId": "1"
+      "userId": "1",
+      "id": "deda"
     },
     {
       "tradeId": "2",
@@ -207,6 +208,71 @@
       "tradeDescription": "주식 배당금",
       "userId": "5",
       "id": "1d08"
+    },
+    {
+      "id": "1744188433145",
+      "tradeId": "1744188433145",
+      "tradeType": "지출",
+      "tradeDate": "2025-04-09",
+      "tradeAmount": 3234342,
+      "tradeDescription": "",
+      "userId": "1",
+      "expenseCategory": "세금",
+      "tradeMethod": "현금"
+    },
+    {
+      "id": "1744244242839",
+      "tradeId": "1744244242839",
+      "tradeType": "지출",
+      "tradeDate": "2025-04-10",
+      "tradeAmount": 1500000,
+      "tradeDescription": "이마트 장보기",
+      "userId": "1",
+      "expenseCategory": "식비",
+      "tradeMethod": "신용카드"
+    },
+    {
+      "id": "1744248228107",
+      "tradeId": "1744248228107",
+      "tradeType": "지출",
+      "tradeDate": "2025-04-10",
+      "tradeAmount": 5000000,
+      "tradeDescription": "전세 집 구함 꺄르륵",
+      "userId": "userStore.userId",
+      "expenseCategory": "주거",
+      "tradeMethod": "계좌이체"
+    },
+    {
+      "id": "1744248340490",
+      "tradeId": "1744248340490",
+      "tradeType": "지출",
+      "tradeDate": "2025-04-10",
+      "tradeAmount": 50000000,
+      "tradeDescription": "집을 계좌이체로 삼",
+      "userId": "3",
+      "expenseCategory": "주거",
+      "tradeMethod": "현금"
+    },
+    {
+      "id": "1744248390846",
+      "tradeId": "1744248390846",
+      "tradeType": "수입",
+      "tradeDate": "2025-04-10",
+      "tradeAmount": 50000,
+      "tradeDescription": "알바 대타갔다옴 ",
+      "userId": "3",
+      "incomeCategory": "급여"
+    },
+    {
+      "id": "1744248442859",
+      "tradeId": "1744248442859",
+      "tradeType": "지출",
+      "tradeDate": "2025-04-10",
+      "tradeAmount": 500000000,
+      "tradeDescription": "돈모으는 여자 ",
+      "userId": "5",
+      "expenseCategory": "저축",
+      "tradeMethod": "현금"
     }
   ],
   "tradeType": [
