+++ src/assets/addTrade.css
@@ -67,7 +67,7 @@ padding: 10px 1rem 0;  display: flex;
   padding: 1rem;
   display: flex;
   flex-direction: column;
-  gap: 2rem;
+  gap: 0.5rem;
   /* 각 항목 간 간격 */
 }
